Fix weapon statues to always give loot item
